Storing Token Transactions
I extended the Token's blueprint to support storing every transaction made (transfers and mints).

This is crucial as users need to know a token's transaction to protect themself from bad actors by verifying everything that goes on in the transactions.

I removed the tokenV2 file that was an extension of the token's blueprint with TransferFrom and Approve support as they are no longer needed; we can already use "X-Forwarded" tags instead.